requirements.txt
Flask==3.1.*
msal==1.*
identity[flask]==0.7.*
requests==2.*
urllib3==2.*
APScheduler==3.10.*
Flask-APScheduler==1.13.*
gunicorn==22.*
python-dotenv==1.*
azure-data-tables==12.*
azure-identity>=1.17
azure-keyvault-secrets>=4.8
PyJWT>=2.8
